♻️ refactor(task): fold the route chrome into TaskDetailSkeleton
One skeleton component serves the route (with NavHeader) and the page /
Portal bodies via the chrome prop, instead of a second wrapper file.
