+++ src/features/AgentTasks/routeMeta.ts
@@ -1,7 +1,7 @@
 import { ListTodoIcon } from 'lucide-react';
 
-import TaskDetailSkeleton from '@/components/Skeleton/TaskDetail';
 import TasksSkeleton from '@/components/Skeleton/Tasks';
+import TaskDetailSkeleton from '@/features/AgentTasks/AgentTaskDetail/TaskDetailSkeleton';
 import { usePublishDynamicRouteMeta } from '@/features/RouteMeta/usePublishDynamicRouteMeta';
 import { matchesRouteWorkspace, useRouteWorkspaceId } from '@/features/RouteMeta/workspaceScope';
 import type { DynamicRouteMetaProps } from '@/spa/router/routeMeta';
